requirements.txt
blinker==1.8.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
flask==3.0.3
fonttools==4.55.3
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.5
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
